psginstion fixed | todo - add paxination on room table

+++ controllers/userController.js
@@ -4,7 +4,9 @@ const cloudinary = require('cloudinary').v2;
 const secretKey ="mysecretKey"
 const bcrypt = require('bcrypt');
 const RoomTypes = require('../models/roomType')
+const Room= require('../models/rooms')
 const fs = require('fs');
+const Rooms = require('../models/rooms');
 
 module.exports.signup=async function(req,res){
     try {
@@ -78,6 +80,8 @@ module.exports.login=async function(req,res){
         
         // Set the token in a cookie
         res.cookie('token', token, { httpOnly: true });
+        res.locals.user = user;
+        // req.session.user = user;
     
        res.redirect('/');
       } catch (error) {
@@ -88,6 +92,7 @@ module.exports.login=async function(req,res){
 
 module.exports.addCategoryPage=async(req,res)=>{
   const userId = req.user.userId
+  const user = await User.findById(userId)
   const category = await RoomTypes.find({owner:userId})
   return res.render('addcategory',{
     title:'Category',
@@ -122,12 +127,29 @@ module.exports.signout = function (req, res) {
 };
 
 
-module.exports.deleteType = async(req,res)=>{
-  const {id} = req.params
-  const type = await RoomTypes.deleteOne({_id:id})
+module.exports.deleteType = async (req, res) => {
+  try {
+    const { id } = req.params;
+    const userId = req.user.userId;
 
-  return res.redirect('back')
-}
+    // Delete the room type
+    const type = await RoomTypes.deleteOne({ _id: id, owner: userId });
+
+    // Delete associated rooms
+    const room = await Rooms.deleteMany({ roomTypeId: id, owner: userId });
+
+    // Check if any documents were deleted
+    if (type.deletedCount === 0) {
+      return res.status(404).json({ message: 'Room type not found' });
+    }
+
+    // Redirect only if the deletions were successful
+    return res.redirect('back');
+  } catch (error) {
+    console.error(error);
+    return res.status(500).json({ message: 'Internal server error' });
+  }
+};
 
 module.exports.myAccount=async(req,res)=>{
   try{
